refactor(move): branch ref code into util file
