+++ src/lib/utils/index.ts
@@ -1,3 +1,4 @@
+import { getBranchChildrenOrParentsFromGit } from "./branch_refs";
 import { checkoutBranch } from "./checkout_branch";
 import { getCommitterDate } from "./committer_date";
 import { detectStagedChanges } from "./detect_staged_changes";
@@ -39,5 +40,6 @@ export {
   preprocessCommand,
   signpostDeprecatedCommands,
   logTip,
+  getBranchChildrenOrParentsFromGit,
   VALIDATION_HELPER_MESSAGE,
 };
