+++ src/wrapper-classes/branch.ts
@@ -1,10 +1,11 @@
-import chalk from "chalk";
 import { execSync } from "child_process";
-import { repoConfig } from "../lib/config";
 import { ExitFailedError } from "../lib/errors";
-import { tracer } from "../lib/telemetry";
-import { getCommitterDate, getTrunk, gpExecSync } from "../lib/utils";
-import { logDebug } from "../lib/utils/splog";
+import {
+  getBranchChildrenOrParentsFromGit,
+  getCommitterDate,
+  getTrunk,
+  gpExecSync,
+} from "../lib/utils";
 import Commit from "./commit";
 
 type TMeta = {
@@ -15,149 +16,6 @@ type TMeta = {
     url: string;
   };
 };
-
-function gitTreeFromRevListOutput(output: string): Record<string, string[]> {
-  const ret: Record<string, string[]> = {};
-  for (const line of output.split("\n")) {
-    if (line.length > 0) {
-      const shas = line.split(" ");
-      ret[shas[0]] = shas.slice(1);
-    }
-  }
-
-  return ret;
-}
-
-function getChildrenRevListGitTree(
-  branchName: string
-): Record<string, string[]> {
-  const otherBranches = Branch.allBranches()
-    .map((b) => b.name)
-    .filter((b) => b !== branchName);
-
-  return gitTreeFromRevListOutput(
-    execSync(
-      // Check that there is a commit behind this branch before getting the full list.
-      `git show-ref ${branchName}~1 && git rev-list --children ${otherBranches.join(
-        " "
-        // If there are no commits behind current, just fetch all refs.
-      )} ^${branchName}~1 || git rev-list --children --all`,
-      {
-        maxBuffer: 1024 * 1024 * 1024,
-      }
-    )
-      .toString()
-      .trim()
-  );
-}
-
-function getParentRevListGitTree(branchName: string): Record<string, string[]> {
-  return gitTreeFromRevListOutput(
-    execSync(`git rev-list --parents ${branchName} ^${getTrunk().name}`, {
-      maxBuffer: 1024 * 1024 * 1024,
-    })
-      .toString()
-      .trim()
-  );
-}
-
-function branchListFromShowRefOutput(output: string): Record<string, string[]> {
-  const ret: Record<string, string[]> = {};
-  const ignorebranches = repoConfig.getIgnoreBranches();
-
-  for (const line of output.split("\n")) {
-    if (line.length > 0) {
-      const parts = line.split(" ");
-      const branchName = parts[1].slice("refs/heads/".length);
-      const branchRef = parts[0];
-      if (!ignorebranches.includes(branchName)) {
-        if (branchRef in ret) {
-          ret[branchRef].push(branchName);
-        } else {
-          ret[branchRef] = [branchName];
-        }
-      }
-    }
-  }
-
-  return ret;
-}
-
-let memoizedBranchList: Record<string, string[]>;
-function getBranchList(opts: {
-  useMemoizedResult?: boolean;
-}): Record<string, string[]> {
-  if (opts.useMemoizedResult && memoizedBranchList !== undefined) {
-    return memoizedBranchList;
-  }
-
-  memoizedBranchList = branchListFromShowRefOutput(
-    execSync("git show-ref --heads", {
-      maxBuffer: 1024 * 1024 * 1024,
-    })
-      .toString()
-      .trim()
-  );
-
-  return memoizedBranchList;
-}
-
-function traverseGitTreeFromCommitUntilBranch(
-  commit: string,
-  gitTree: Record<string, string[]>,
-  branchList: Record<string, string[]>,
-  n: number
-): {
-  branches: Set<string>;
-  shortCircuitedDueToMaxDepth?: boolean;
-} {
-  // Skip the first iteration b/c that is the CURRENT branch
-  if (n > 0 && commit in branchList) {
-    return {
-      branches: new Set(branchList[commit]),
-    };
-  }
-
-  // Limit the seach
-  const maxBranchLength = repoConfig.getMaxBranchLength();
-  if (n > maxBranchLength) {
-    return {
-      branches: new Set(),
-      shortCircuitedDueToMaxDepth: true,
-    };
-  }
-
-  if (!gitTree[commit] || gitTree[commit].length == 0) {
-    return {
-      branches: new Set(),
-    };
-  }
-
-  const commitsMatchingBranches = new Set<string>();
-  let shortCircuitedDueToMaxDepth = undefined;
-  for (const neighborCommit of gitTree[commit]) {
-    const results = traverseGitTreeFromCommitUntilBranch(
-      neighborCommit,
-      gitTree,
-      branchList,
-      n + 1
-    );
-
-    const branches = results.branches;
-    shortCircuitedDueToMaxDepth =
-      results.shortCircuitedDueToMaxDepth || shortCircuitedDueToMaxDepth;
-
-    if (branches.size !== 0) {
-      branches.forEach((commit) => {
-        commitsMatchingBranches.add(commit);
-      });
-    }
-  }
-  return {
-    branches: commitsMatchingBranches,
-    shortCircuitedDueToMaxDepth: shortCircuitedDueToMaxDepth,
-  };
-}
 
 type TBranchFilters = {
   useMemoizedResults?: boolean;
@@ -241,6 +99,7 @@ export default class Branch {
   }
 
   getParentFromMeta(): Branch | undefined {
+    console.log(`getting meta parent ${this.name}`);
     if (this.name === getTrunk().name) {
       return undefined;
     }
@@ -274,6 +133,7 @@ export default class Branch {
   }
 
   public getChildrenFromMeta(): Branch[] {
+    console.log(`getting meta children ${this.name}`);
     const children = Branch.allBranches().filter(
       (b) => b.getMeta()?.parentBranchName === this.name
     );
@@ -533,13 +393,15 @@ export default class Branch {
   }
 
   public getChildrenFromGit(): Branch[] {
-    return this.getChildrenOrParents({
+    console.log(`get children from git ${this.name}`);
+    return getBranchChildrenOrParentsFromGit(this, {
       direction: "CHILDREN",
       useMemoizedResults: this.shouldUseMemoizedResults,
     });
   }
 
   public getParentsFromGit(): Branch[] {
+    console.log(`get parents from git ${this.name}`);
     if (
       // Current branch is trunk
       this.name === getTrunk().name
@@ -549,7 +411,7 @@ export default class Branch {
     } else if (this.pointsToSameCommitAs(getTrunk())) {
       return [getTrunk()];
     }
-    return this.getChildrenOrParents({
+    return getBranchChildrenOrParentsFromGit(this, {
       direction: "PARENTS",
       useMemoizedResults: this.shouldUseMemoizedResults,
     });
@@ -557,59 +419,6 @@ export default class Branch {
 
   private pointsToSameCommitAs(branch: Branch): boolean {
     return !!this.branchesWithSameCommit().find((b) => b.name === branch.name);
-  }
-
-  private getChildrenOrParents(opt: {
-    direction: "CHILDREN" | "PARENTS";
-    useMemoizedResults?: boolean;
-  }): Branch[] {
-    const direction = opt.direction;
-    const useMemoizedResults = opt.useMemoizedResults ?? false;
-    return tracer.spanSync(
-      {
-        name: "function",
-        resource: "branch.getChildrenOrParents",
-        meta: { direction: direction },
-      },
-      () => {
-        const gitTree =
-          direction === "CHILDREN"
-            ? getChildrenRevListGitTree(this.name)
-            : getParentRevListGitTree(this.name);
-
-        const headSha = execSync(`git rev-parse ${this.name}`)
-          .toString()
-          .trim();
-
-        const childrenOrParents = traverseGitTreeFromCommitUntilBranch(
-          headSha,
-          gitTree,
-          getBranchList({ useMemoizedResult: useMemoizedResults }),
-          0
-        );
-
-        if (childrenOrParents.shortCircuitedDueToMaxDepth) {
-          logDebug(
-            `${chalk.magenta(
-              `Potential missing branch ${direction.toLocaleLowerCase()}:`
-            )} Short-circuited search for branch ${chalk.bold(
-              this.name
-            )}'s ${direction.toLocaleLowerCase()} due to Graphite 'max-branch-length' setting. (Your Graphite CLI is currently configured to search a max of <${repoConfig.getMaxBranchLength()}> commits away from a branch's tip.) If this is causing an incorrect result (e.g. you know that ${
-              this.name
-            } has ${direction.toLocaleLowerCase()} ${
-              repoConfig.getMaxBranchLength() + 1
-            } commits away), please adjust the setting using \`gt repo max-branch-length\`.`
-          );
-        }
-
-        return Array.from(childrenOrParents.branches).map((name) => {
-          const branch = new Branch(name);
-          return this.shouldUseMemoizedResults
-            ? branch.useMemoizedResults()
-            : branch;
-        });
-      }
-    );
   }
 
   public setPRInfo(prInfo: { number: number; url: string }): void {
